tests/wpt/web-platform-tests/css/mediaqueries/prefers-reduced-data.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/mediaqueries-5/#prefers-reduced-data" />
<script type="text/javascript" src="/resources/testharness.js"></script>
<script type="text/javascript" src="/resources/testharnessreport.js"></script>

<script type="text/javascript" src="resources/matchmedia-utils.js"></script>
<script>
query_should_be_css_parseable("(prefers-reduced-data)");
query_should_be_css_parseable("(prefers-reduced-data: no-preference)");
query_should_be_css_parseable("(prefers-reduced-data: reduce)");

query_should_not_be_css_parseable("(prefers-reduced-data: 0)");
query_should_not_be_css_parseable("(prefers-reduced-data: none)");
query_should_not_be_css_parseable("(prefers-reduced-data: 10px)");
query_should_not_be_css_parseable("(prefers-reduced-data: no-preference reduce)");
query_should_not_be_css_parseable("(prefers-reduced-data: reduced)");
query_should_not_be_css_parseable("(prefers-reduced-data: no-preference/reduce)");

query_should_be_js_parseable("(prefers-reduced-data)");
query_should_be_js_parseable("(prefers-reduced-data: no-preference)");
query_should_be_js_parseable("(prefers-reduced-data: reduce)");

query_should_not_be_js_parseable("(prefers-reduced-data: 0)");
query_should_not_be_js_parseable("(prefers-reduced-data: none)");
query_should_not_be_js_parseable("(prefers-reduced-data: 10px)");
query_should_not_be_js_parseable("(prefers-reduced-data: no-preference reduce)");
query_should_not_be_js_parseable("(prefers-reduced-data: reduced)");
query_should_not_be_js_parseable("(prefers-reduced-data: no-preference/reduce)");

test(() => {
  // What this is saying is that 'no-preference' is not the default, so
  // irregardless of the current OS settings, (prefers-reduced-data).matches
  // should not be equivalent to (prefers-reduced-data: no-preference).matches.
  let booleanContext = window.matchMedia("(prefers-reduced-data)");
  let noPreference = window.matchMedia("(prefers-reduced-data: no-preference)");
  assert_equals(booleanContext.matches, !noPreference.matches);
}, "Check that no-preference evaluates to false in the boolean context");
</script>
